+++ CHANGELOG.md
@@ -34,7 +34,9 @@
   - Add validation on tx_id search to force a valid UUID string to be entered
   - FTP files write to memory instead of writing temporary files
   - Add pagination in store data
+  - Remove /clear endpoint
   - Create database tables in development mode
+
 
 ### 1.7.0 2017-11-30
   - Deploying new console